refactor(ui): extract health/render/stats.ts — stats grid + meta line
Move renderStats out of health.ts into health/render/stats.ts. The
module owns the project-filtered stat grid (savings, injected tokens,
reduction percent, work investment, active memories, embedding/tag
coverage, raw event counters) plus the DB path + size meta line. The
barrel re-exports renderStats so existing call sites stay on the same
import path.
